database seeded successfully with all information! hooray! also finished models

+++ nyc_schools_app/db/migrate/20140708231330_create_districts.rb
@@ -2,8 +2,8 @@ class CreateDistricts < ActiveRecord::Migration
   def change
     create_table :districts do |t|
       t.string :district_title
-      t.integer :avg_attendance
-      t.float :avg_enrollment
+      t.float :avg_attendance
+      t.integer :avg_enrollment
       t.integer :math_num_tested
       t.float :"math_pct_lvl1"
       t.float :"math_pct_lvl2"
